requirements.txt
# Core
sqlmodel>=0.0.21
asyncpg>=0.29.0
alembic>=1.13.0
pydantic-settings>=2.0.0
greenlet>=3.0.0

# Scraper
httpx>=0.27.0
beautifulsoup4>=4.12.0
aiohttp>=3.9.0

# Worker
arq>=0.25.0
redis>=5.0.0
tenacity>=8.3.0

# API
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
